Implement regression test feature
Actually allows to check difference of performance between two build
result list

Also implements n_runs parameter, to do an average of multiple runs of
the same test
While unacceptable_n_runs is an amount of runs to do once a variable is
rejected. Before saying that performances are too much different, we do
for example 3 more runs and check once again the average, but only on
half the values, the one central to the median to exclude abnormal
values.

This is because I saw that once a few times I get 27G on some test
instead of 29G. This may be due to external hardware or CPU throttle, or
impact from the environment. So this seems to be a good behaviour.