Multiple commits
If --output is given without argument, it will put CSVs with the graph.
Same for --statistics

Var_repeat will ensure that given variable has a point for ervery time,
duplicating it if needed. It is usefull to ensure one data point for
each graph. Eg if run 1 has value for time 1 and 3 and run 2 has value
for 2 only, you don't want to see a global variation of 1 to 2 to 3, has
    in reality experiment were stable independently