Experimental design
Allow to define covariables and then use experimental design matrix to
propose values in a range

Take the following example:
A=[1-10]
B=[1-10]

This would lead to 100 tests, for (A=1, B=1), (A=1, B=2) ... (A=10,
B=10)

Now a covariable such as this will make variables move together:

{
    A=[1-10]
    B=[1-10]
}

Will lead to 10 tests, (A=1,B=1), (A=2, B=2) ... (A=10, B=10)

The new parameter for ranges allows to draw values from a matrix of
equal distances (see J. Santiago, M. Claeys-Bruno, and M. Sergent. Construction of Space-Filling Designs using WSP Algorithm for High Dimensional Spaces. Chemometrics and Intelligent Laboratory Systems, 113, 2012.), such as this:

{
    A=[1-1000#]
    B=[1-10000#]
    C=[1-2000000#]
}
Definitely, no one wants to execute 2M*10K*1K=10^13 runs. A range with
no step (nothing after the #) will take 95 values out of the range, in a
way that the distance between all points in A, B and C is optimal (as
defined in WSP). Also, the value accross multiple runs will always be
the same, so the experiments are reproducibles.

Now that will still have 4 variables with different values, and the
plot would probably not look like what you'd want. So the configuration
option var_aggregate now accepts lists:

%config
var_aggregate={A+B+C:all}

All points will be combined as a single run and will just be considered
as "variance" of the other variables.