Enable "local" repository
If a repository is defined with the methode "local", then the "url" will
be used as a local path.

This is useful to specify a software you're working on locally. One can
directly define a "local" repository on te commandline instead of the
.repo file using "local/path/to/file:MySoftware".

In conjuction with nfs=0 in the cluster file, this will rsync your
"local" folder to the server before running the test. Ie this is a full
cmdline example without configuration file:

npf-run.py local/path/to/folder/ --test integration/localcopy.npf --cluster dut=user@10.0.0.1:/tmp/npf/,nfs=0