Add the ifeq-VAR=VALUE parameter to scripts
If the script has ifeq-VAR=VALUE parameter, it will only run if the
variable has that value.

This helps with a particular case of a variables being used to test
various programs instead of a tag.

Fix also the fact that multiple dataset would use the same panda CSV
file