Refactoring
Split huge files in a more reasonable hierarchy.

Change name of handlers. npf-compare becomes npf, as it's the main use
case.

npf-run becomes npf-regress, as the functionality it provides is to look
for the resultso f multiple tests and go through git history

npf-watch stays npf-watch (track a git repo and send a mail if there's a
regression)